locations and no more pages dir
